Permissions: Fix origin mismatch with special URL schemes
Following the Chromium 140 integration, there was a mismatch
between the origin that was stored in the permanent store, and
the one Chromium queried when setting a permission for a custom URL
scheme with no base origin (e.g. "qrc://filename.html"). Specifically,
we would store "qrc://", but Chromium would look up "qrc:". Since
the lookup is text-based, the permission query would always return ASK.

This commit adds conversions to url::Origin before storing
and retrieving the permission, which makes the results equal again.

Pick-to: 6.11
Change-Id: I929c63d4395ec4382f1f93590ca3528b7562e2e6
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>