Pass Nodejs_EXECUTABLE to GN build
The Chromium build system used to find nodejs separately by running
'which nodejs' and 'which node'. This led to issues when the executable
found differed from the one found by qtwebengine's CMake configuration.
Now we pass the path directly to the Chromium build via an environment
variable so it can be read in third_party/node/node.py.

Pick-to: 6.8 6.9 6.10 6.11
Fixes: QTBUG-142667
Task-number: QTBUG-124576
Change-Id: Idf0d040aed8f29a8d4aa6a1269c980de885863ef
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>